- Added advanced database components - Some logic changes

+++ src/main/java/com/group25/ebnisina/managedoctors/dto/DoctorDTO.java
@@ -3,6 +3,8 @@ package com.group25.ebnisina.managedoctors.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class DoctorDTO {
@@ -15,5 +17,5 @@ public class DoctorDTO {
 
     private String last_name;
 
-    private Integer rating;
+    private BigDecimal rating;
 }
